refs #18 refactor utils directory structure
