+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@tyranno269/tatsumaki",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "description": "Automatically generate TypeSpec models from Rails database schema",
   "type": "module",
   "main": "dist/index.mjs",
