+++ src/discovery/modelFinder.test.ts
@@ -12,7 +12,7 @@ vi.mock("fast-glob", () => ({
   default: vi.fn(),
 }));
 
-vi.mock("./enumParser.js", () => ({
+vi.mock("../parsing/enumParser.js", () => ({
   parseRailsEnums: vi.fn(),
   getModelNameFromPath: vi.fn(),
 }));
@@ -21,7 +21,7 @@ describe("parseRelevantRailsEnums", () => {
   test("filters models based on table names", async () => {
     const { readFileSync, existsSync } = await import("fs");
     const fg = (await import("fast-glob")).default;
-    const { parseRailsEnums, getModelNameFromPath } = await import("./enumParser.js");
+    const { parseRailsEnums, getModelNameFromPath } = await import("../parsing/enumParser.js");
 
     // Setup mocks
     vi.mocked(fg).mockResolvedValue([
@@ -33,14 +33,23 @@ describe("parseRelevantRailsEnums", () => {
     vi.mocked(existsSync).mockReturnValue(true);
     vi.mocked(readFileSync).mockReturnValue("mock content");
 
-    vi.mocked(getModelNameFromPath)
-      .mockReturnValueOnce("Company")
-      .mockReturnValueOnce("Prefecture")
-      .mockReturnValueOnce("User");
+    // Mock the imported functions
+    (getModelNameFromPath as ReturnType<typeof vi.fn>).mockImplementation((path: string) => {
+      if (path.includes("company.rb")) return "Company";
+      if (path.includes("prefecture.rb")) return "Prefecture";
+      if (path.includes("user.rb")) return "User";
+      return "";
+    });
 
-    vi.mocked(parseRailsEnums)
-      .mockReturnValueOnce([{ fieldName: "status", values: ["active"], modelName: "Company" }])
-      .mockReturnValueOnce([{ fieldName: "region", values: ["kanto"], modelName: "User" }]);
+    (parseRailsEnums as ReturnType<typeof vi.fn>).mockImplementation(
+      (content: string, modelName: string) => {
+        if (modelName === "Company")
+          return [{ fieldName: "status", values: ["active"], modelName: "Company" }];
+        if (modelName === "User")
+          return [{ fieldName: "region", values: ["kanto"], modelName: "User" }];
+        return [];
+      },
+    );
 
     // Only companies and users tables exist in schema.rb
     const tableNames = ["companies", "users"];
